fix bug in purgecss extractor
I was using part of this purgecss config for a project of mine, and i kept getting an `Error: The extractor has failed to extract the selectors.` error. Turns out the extractor returns `null` when no selectors are found, which causes purgecss to fail. Returning an empty array fixes this